add full uuid macro
