+++ Cargo.toml
@@ -9,15 +9,13 @@ name = "uuid_macro"
 repository = "https://github.com/uuid-rs/uuid_macro"
 version = "0.1.0"
 
-[lib]
-proc-macro = true
 
-[dependencies.syn]
-version = "0.13"
+[dependencies]
+uuid_macro_impl = { path = "./uuid_macro_impl"}
+proc-macro-hack = { version = "0.4"}
+uuid = { git = "https://github.com/mokeyish/uuid.git", branch = "master" }
+#uuid = { git = "https://github.com/mokeyish/uuid.git", branch = "master", features = ["nightly"] }
 
-[dependencies.quote]
-version = "0.5"
-
-[dependencies.proc-macro-hack]
-version = "0.4"
-
+[features]
+default = []
+nightly = []
